+++ writer_test.go
@@ -33,8 +33,8 @@ func scanParquetColumns(col *parquet.Column) error {
 
 		for pages.Next() {
 			numValues := pages.NumValues()
-			repetitions := make([]int32, numValues)
-			definitions := make([]int32, numValues)
+			repetitions := make([]int8, numValues)
+			definitions := make([]int8, numValues)
 
 			var n int
 			var err error
